requirements.txt
dj-rest-kit==1.1.1
Django==5.0.1
django-filter==23.5
django-json-widget==1.1.1
django-model-utils==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
django-modeltranslation==0.18.11
django-environ==0.11.2
django-cors-headers==4.3.1
Pillow==10.2.0
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
django-tinymce==3.6.1
djangorestframework-simplejwt==5.3.1